+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+## v0.7.8 - 2023-06-07
+
 ### Added
 - Add a new command `docker-build` to build a docker image using a Dockerfile without using an Earthfile, locally or on a satellite.